Big Update for First Presentation
Updated Door Prefars to properly work and replaced all current doors with them. Setup a placeholder Item prefab and dropped a few into the current scenes. Modified the testing SceneSwapTest script as well as "fixing" a few of the other legacy scripts so they don't popup errors in the console all the time. Built and ran a new build.